dom/network/tests/test_network_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Network API</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Network Information API **/
function test() {
  ok('connection' in navigator, "navigator.connection should exist");

  ok(navigator.connection, "navigator.connection returns an object");

  ok(navigator.connection instanceof EventTarget,
     "navigator.connection is a EventTarget object");

  ok('type' in navigator.connection,
     "type should be a Connection attribute");
  is(navigator.connection.type, "none",
     "By default connection.type equals to none");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({'set': [["dom.netinfo.enabled", true]]}, test);

</script>
</pre>
</body>
</html>
